+++ app/templates/navbar.html
@@ -10,17 +10,17 @@
             <ul class="navbar-nav me-auto mb-2 mb-lg-0">
                 {% if current_user.is_authenticated %}
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ url_for('dashboard')}}">
+                        <a class="nav-link active" aria-current="page" href="{{ url_for('user_dashboard')}}">
                             Dashboard
                         </a>
                     </li>  
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ url_for('logout')}}">
+                        <a class="nav-link active" aria-current="page" href="{{ url_for('user_logout')}}">
                             Logout
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ url_for('add_post')}}">
+                        <a class="nav-link active" aria-current="page" href="{{ url_for('post_add')}}">
                             Bericht plaatsen
                         </a>
                     </li>
@@ -30,18 +30,18 @@
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ url_for('add_user')}}">
+                        <a class="nav-link active" aria-current="page" href="{{ url_for('users')}}">
                             Gebruikers
                         </a>
                     </li>
                 {% else %}
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ url_for('login')}}">
+                        <a class="nav-link active" aria-current="page" href="{{ url_for('user_login')}}">
                             Login
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ url_for('add_user')}}">
+                        <a class="nav-link active" aria-current="page" href="{{ url_for('user_add')}}">
                             Registreren
                         </a>
                     </li>
