+++ app/templates/post.html
@@ -25,10 +25,10 @@
 <a href="{{ url_for('posts')}}" class="btn btn-success btn-sm">
     Terug
 </a>
-<a href="{{ url_for('edit_post', id=post.id)}}" class="btn btn-outline-success btn-sm">
+<a href="{{ url_for('post_edit', id=post.id)}}" class="btn btn-outline-success btn-sm">
     Bewerken
 </a>
-<a href="{{ url_for('delete_post', id=post.id)}}" class="btn btn-outline-danger btn-sm">
+<a href="{{ url_for('post_delete', id=post.id)}}" class="btn btn-outline-danger btn-sm">
     Verwijderen
 </a>
