+++ app/templates/posts.html
@@ -23,13 +23,13 @@
         {{ post.content }}
         <br/>
         <br/>
-        <a href="{{ url_for('blog_post', id=post.id)}} " class = "btn btn-success btn-sm">
+        <a href="{{ url_for('post', id=post.id)}} " class = "btn btn-success btn-sm">
             Ga naar bericht
         </a>
-        <a href="{{ url_for('edit_post', id=post.id)}} " class = "btn btn-outline-success btn-sm">
+        <a href="{{ url_for('post_edit', id=post.id)}} " class = "btn btn-outline-success btn-sm">
             Bewerken
         </a>
-        <a href="{{ url_for('delete_post', id=post.id)}}" class="btn btn-outline-danger btn-sm">
+        <a href="{{ url_for('post_delete', id=post.id)}}" class="btn btn-outline-danger btn-sm">
             Verwijderen
         </a>
     </div>
