--- app/templates/user.html
@@ -1,4 +1,0 @@
-{% extends 'base.html'%}
-{% block content%}
-    <h1>Hoi {{ name_|capitalize }}</h1>
-{% endblock %}
