+++ app/templates/user_dashboard.html
@@ -28,9 +28,9 @@
                 <strong>Favoriete kleur: </strong>{{ current_user.favo_kl }}</br>
                 <strong>Lid sinds: </strong>{{ current_user.dates }}</br>    
             </p>
-            <a href="{{ url_for('update', id=current_user.id)}}" class="btn btn-succes btn-sm">Bewerken</a>
-            <a href="{{ url_for('logout') }}" class="btn btn-warning btn-sm">Uitloggen</a>
-            <a href="{{ url_for('delete', id=current_user.id)}}" class="btn btn-danger btn-sm">Profiel verwijderen</a>
+            <a href="{{ url_for('user_update', id=current_user.id)}}" class="btn btn-success btn-sm">Bewerken</a>
+            <a href="{{ url_for('user_logout') }}" class="btn btn-warning btn-sm">Uitloggen</a>
+            <a href="{{ url_for('user_delete', id=current_user.id)}}" class="btn btn-danger btn-sm">Profiel verwijderen</a>
         </div>
     </div>
     <br/><br/>
