+++ app/templates/users.html
@@ -1,0 +1,34 @@
+{% extends 'base.html'%}
+
+{% block content%}
+
+<table class="table table-hover caption-top">
+    <caption>Gebruikers</caption>
+    <thead>
+      <tr>
+        <th scope="col">#</th>
+        <th scope="col">Naam</th>
+        <th scope="col">Gebruikersnaam</th>
+        <th scope="col">Email</th>
+        <th scope="col">Favoriete kleur</th>
+        <th scope="col">Verwijderen</th>
+      </tr>
+    </thead>
+    <tbody>
+        {% for gebruiker in gebruikers %}
+            <tr>
+                <th scope="row">{{ gebruiker.id }}</th>
+                <td>{{ gebruiker.name }}</td>
+                <td>{{ gebruiker.username }}</td>
+                <td>{{ gebruiker.email }}</td>
+                <td>{{ gebruiker.favo_kl }}</td>
+                <td colspan="2">
+                    <a href="{{ url_for('user_update', id=gebruiker.id)}}" class="btn btn-outline-warning btn-sm"> Edit </a>
+                    <a href="{{ url_for('user_delete', id=gebruiker.id)}}" class="btn btn-outline-danger btn-sm"> Delete </a>
+                </td>
+            </tr>
+        {% endfor %}
+    </tbody>
+</table>
+ 
+{% endblock %}
